Python_SQL/requirements.txt
mysql-connector-python python-dotenv